prevent user go to admin panel

+++ app/Http/Controllers/DocumentController.php
@@ -21,27 +21,33 @@ class DocumentController extends Controller
 
     public function index()
     {
-        // get data from document table
-        $doc = DB::table('documents')
-            ->leftJoin('document_chapters', 'documents.id', '=', 'document_chapters.document_id')
-            ->select('documents.*', DB::raw('COUNT(document_chapters.document_id) as document_chapter_count'),  DB::raw('TRUNCATE(SUM(document_chapters.cost_of_translate),2) as cost_of_translate')) //truncate to standarize 2 digit after comma
-            ->where('document_chapters.status', '=', '10') // show the doc with chapter
-            // show doc with the various document chapter status
-            ->orWhere('document_chapters.status', '=', '0') // created
-            ->orWhere('document_chapters.status', '=', '1') // ongoing
-            ->orWhere('document_chapters.status', '=', '2') // submitted
-            ->orWhere('document_chapters.status', '=', '3') // rejected
-            ->orWhereNull('document_chapters.status')   //show new doc without chapter
-            ->groupBy('documents.id')
-            ->get();
+        $role = auth()->user()->role;
+        // SECURE ADMIN ROUTING
+        if ($role == 1) {
+            // get data from document table
+            $doc = DB::table('documents')
+                ->leftJoin('document_chapters', 'documents.id', '=', 'document_chapters.document_id')
+                ->select('documents.*', DB::raw('COUNT(document_chapters.document_id) as document_chapter_count'),  DB::raw('TRUNCATE(SUM(document_chapters.cost_of_translate),2) as cost_of_translate')) //truncate to standarize 2 digit after comma
+                ->where('document_chapters.status', '=', '10') // show the doc with chapter
+                // show doc with the various document chapter status
+                ->orWhere('document_chapters.status', '=', '0') // created
+                ->orWhere('document_chapters.status', '=', '1') // ongoing
+                ->orWhere('document_chapters.status', '=', '2') // submitted
+                ->orWhere('document_chapters.status', '=', '3') // rejected
+                ->orWhereNull('document_chapters.status')   //show new doc without chapter
+                ->groupBy('documents.id')
+                ->get();
 
-        // dd($doc);
+            // dd($doc);
 
-        return view('pages.admin.document.index')->with(
-            [
-                'doc' => $doc
-            ]
-        );
+            return view('pages.admin.document.index')->with(
+                [
+                    'doc' => $doc
+                ]
+            );
+        } else {
+            return redirect()->back();
+        }
     }
 
     /**
@@ -51,7 +57,14 @@ class DocumentController extends Controller
      */
     public function create()
     {
-        return view('pages.admin.document.create');
+        $role = auth()->user()->role;
+
+        // SECURE ADMIN ROUTING
+        if ($role == 1) {
+            return view('pages.admin.document.create');
+        } else {
+            return redirect()->back();
+        }
     }
 
     /**
@@ -62,19 +75,26 @@ class DocumentController extends Controller
      */
     public function store(Request $request)
     {
-        $validatedData = $request->validate([
-            'id_title' => '',
-            'ch_title' => 'required',
-            'number_chapter' => 'required|int',
-            // 'number_chapter_done' => 'required|int',
-            // 'cost_of_translate' => 'required|int',
-            // 'note' => 'required',
-        ]);
-        // dd('lolos');
+        $role = auth()->user()->role;
 
-        Document::create($validatedData);
-        // $request->session()->flash('success', 'Registration User Successfully');
-        return redirect()->route('documents.index');
+        // SECURE ADMIN ROUTING
+        if ($role == 1) {
+            $validatedData = $request->validate([
+                'id_title' => '',
+                'ch_title' => 'required',
+                'number_chapter' => 'required|int',
+                // 'number_chapter_done' => 'required|int',
+                // 'cost_of_translate' => 'required|int',
+                // 'note' => 'required',
+            ]);
+            // dd('lolos');
+
+            Document::create($validatedData);
+            // $request->session()->flash('success', 'Registration User Successfully');
+            return redirect()->route('documents.index');
+        } else {
+            return redirect()->back();
+        }
     }
 
     /**
@@ -96,13 +116,20 @@ class DocumentController extends Controller
      */
     public function edit($id)
     {
-        $doc = Document::findOrFail($id);
+        $role = auth()->user()->role;
 
-        return view('pages.admin.document.edit')->with(
-            [
-                'doc' => $doc
-            ]
-        );
+        // SECURE ADMIN ROUTING
+        if ($role == 1) {
+            $doc = Document::findOrFail($id);
+
+            return view('pages.admin.document.edit')->with(
+                [
+                    'doc' => $doc
+                ]
+            );
+        } else {
+            return redirect()->back();
+        }
     }
 
     /**
@@ -114,16 +141,23 @@ class DocumentController extends Controller
      */
     public function update(Request $request, $id)
     {
-        $validatedData = $request->validate([
-            'id_title' => 'required',
-            'ch_title' => 'required',
-            'number_chapter' => 'required|int',
-        ]);
+        $role = auth()->user()->role;
 
-        $doc = Document::findOrFail($id);
-        $doc->update($validatedData);
-        // $request->session()->flash('success-edit', 'Edit User Successfully');
-        return redirect()->route('documents.index');
+        // SECURE ADMIN ROUTING
+        if ($role == 1) {
+            $validatedData = $request->validate([
+                'id_title' => 'required',
+                'ch_title' => 'required',
+                'number_chapter' => 'required|int',
+            ]);
+
+            $doc = Document::findOrFail($id);
+            $doc->update($validatedData);
+            // $request->session()->flash('success-edit', 'Edit User Successfully');
+            return redirect()->route('documents.index');
+        } else {
+            return redirect()->back();
+        }
     }
 
     /**
@@ -134,12 +168,19 @@ class DocumentController extends Controller
      */
     public function destroy($id)
     {
-        $doc = Document::findOrFail($id);
-        $doc->delete();
+        $role = auth()->user()->role;
 
-        // delete document and delete chapters
-        DocumentChapter::where('document_id', $id)->delete();
+        // SECURE ADMIN ROUTING
+        if ($role == 1) {
+            $doc = Document::findOrFail($id);
+            $doc->delete();
 
-        return redirect()->route('documents.index');
+            // delete document and delete chapters
+            DocumentChapter::where('document_id', $id)->delete();
+
+            return redirect()->route('documents.index');
+        } else {
+            return redirect()->back();
+        }
     }
 }
